+++ src/Opulence/Applications/Tasks/Dispatchers/IDispatcher.php
@@ -10,21 +10,10 @@ namespace Opulence\Applications\Tasks\Dispatchers;
 
 /**
  * Defines the interface for task dispatchers to implement
+ *
+ * @deprecated since 1.0.0-beta4
  */
-interface IDispatcher
+interface IDispatcher extends ITaskDispatcher
 {
-    /**
-     * Dispatches all events of a particular type
-     *
-     * @param string $taskType The type of tasks to dispatch
-     */
-    public function dispatch(string $taskType);
-
-    /**
-     * Registers a task of a certain type
-     *
-     * @param string $taskType The type of task being registered
-     * @param callable $task The task to run
-     */
-    public function registerTask(string $taskType, callable $task);
+    // Don't do anything
 }