+++ src/Opulence/Bootstrappers/ApplicationBinder.php
@@ -8,10 +8,10 @@
  */
 namespace Opulence\Bootstrappers;
 
-use Opulence\Applications\Tasks\Dispatchers\IDispatcher as ITaskDispatcher;
+use Opulence\Applications\Tasks\Dispatchers\ITaskDispatcher as ITaskDispatcher;
 use Opulence\Applications\Tasks\TaskTypes;
 use Opulence\Bootstrappers\Caching\ICache;
-use Opulence\Bootstrappers\Dispatchers\IDispatcher as IBootstrapperDispatcher;
+use Opulence\Bootstrappers\Dispatchers\IBootstrapperDispatcher as IBootstrapperDispatcher;
 
 /**
  * Defines the class that binds the bootstrapper library to the application