+++ src/Opulence/Cryptography/Utilities/Strings.php
@@ -11,7 +11,7 @@ namespace Opulence\Cryptography\Utilities;
 /**
  * Defines some string utilities
  *
- * @deprecated since version 1.0.0-beta3
+ * @deprecated since 1.0.0-beta3
  */
 class Strings
 {