+++ tests/src/Opulence/Framework/Http/Testing/PhpUnit/Mocks/IntegrationTestCase.php
@@ -9,10 +9,10 @@
 namespace Opulence\Tests\Framework\Http\Testing\PhpUnit\Mocks;
 
 use Opulence\Applications\Application;
-use Opulence\Applications\Tasks\Dispatchers\Dispatcher as TaskDispatcher;
+use Opulence\Applications\Tasks\Dispatchers\TaskDispatcher;
 use Opulence\Applications\Tasks\TaskTypes;
 use Opulence\Bootstrappers\BootstrapperRegistry;
-use Opulence\Bootstrappers\Dispatchers\Dispatcher;
+use Opulence\Bootstrappers\Dispatchers\BootstrapperDispatcher;
 use Opulence\Bootstrappers\Paths;
 use Opulence\Debug\Exceptions\Handlers\ExceptionHandler;
 use Opulence\Debug\Exceptions\Handlers\IExceptionHandler;
@@ -88,7 +88,7 @@ class IntegrationTestCase extends BaseIntegrationTestCase
 
         // Setup the bootstrappers
         $bootstrapperRegistry = new BootstrapperRegistry($paths, $this->environment);
-        $bootstrapperDispatcher = new Dispatcher($taskDispatcher, $this->container);
+        $bootstrapperDispatcher = new BootstrapperDispatcher($taskDispatcher, $this->container);
         $bootstrapperRegistry->registerEagerBootstrapper(self::$bootstrappers);
         $taskDispatcher->registerTask(
             TaskTypes::PRE_START,