+++ tests/src/Opulence/Routing/RouterTest.php
@@ -13,7 +13,7 @@ use Opulence\Http\HttpException;
 use Opulence\Http\Requests\Request;
 use Opulence\Http\Requests\RequestMethods;
 use Opulence\Http\Responses\Response;
-use Opulence\Routing\Dispatchers\Dispatcher;
+use Opulence\Routing\Dispatchers\RouteDispatcher;
 use Opulence\Routing\Dispatchers\IDependencyResolver;
 use Opulence\Routing\Dispatchers\MiddlewarePipeline;
 use Opulence\Routing\Routes\Compilers\Compiler;
@@ -66,7 +66,7 @@ class RouterTest extends \PHPUnit\Framework\TestCase
         $this->parser = new Parser();
         $this->compiler = new Compiler($routeMatchers);
         $this->router = new Router(
-            new Dispatcher($dependencyResolver, new MiddlewarePipeline()), 
+            new RouteDispatcher($dependencyResolver, new MiddlewarePipeline()),
             $this->compiler,
             $this->parser
         );